+++ suppaftp/src/async_ftp/mod.rs
@@ -723,10 +723,10 @@ where
     }
 
     /// Execute a command on the server and return the response
-    pub async fn site(&mut self, command: impl ToString) -> FtpResult<()> {
+    pub async fn site(&mut self, command: impl ToString) -> FtpResult<Response> {
         debug!("Sending SITE command: {}", command.to_string());
         self.perform(Command::Site(command.to_string())).await?;
-        self.read_response(Status::CommandOk).await.map(|_| ())
+        self.read_response(Status::CommandOk).await
     }
 
     /// Perform custom command
@@ -734,11 +734,11 @@ where
         &mut self,
         command: impl ToString,
         expected_code: &[Status],
-    ) -> FtpResult<()> {
+    ) -> FtpResult<Response> {
         let command = command.to_string();
         debug!("Sending custom command: {}", command);
         self.perform(Command::Custom(command)).await?;
-        self.read_response_in(expected_code).await.map(|_| ())
+        self.read_response_in(expected_code).await
     }
 
     // -- private